Reduce surface area of flags by defining narrower interface
